fixes federation-request tests
