+++ packages/homeserver/src/homeserver.module.ts
@@ -15,6 +15,7 @@ import { swagger } from '@elysiajs/swagger';
 import type { Emitter } from '@rocket.chat/emitter';
 import Elysia from 'elysia';
 import { invitePlugin } from './controllers/federation/invite.controller';
+import { mediaPlugin } from './controllers/federation/media.controller';
 import { profilesPlugin } from './controllers/federation/profiles.controller';
 import { roomPlugin } from './controllers/federation/rooms.controller';
 import { sendJoinPlugin } from './controllers/federation/send-join.controller';
@@ -27,7 +28,6 @@ import { internalMessagePlugin } from './controllers/internal/message.controller
 import { pingPlugin } from './controllers/internal/ping.controller';
 import { internalRoomPlugin } from './controllers/internal/room.controller';
 import { serverKeyPlugin } from './controllers/key/server.controller';
-import { mediaPlugin } from './controllers/media.controller';
 import { wellKnownPlugin } from './controllers/well-known/well-known.controller';
 
 export type { HomeserverEventSignatures };
@@ -122,8 +122,7 @@ export async function setup(options?: HomeserverSetupOptions) {
 		.use(internalRoomPlugin)
 		.use(serverKeyPlugin)
 		.use(wellKnownPlugin)
-		.use(roomPlugin)
-		.use(mediaPlugin);
+		.use(roomPlugin);
 
 	return { app, container };
 }
